Add files via upload
The holes placement are according to the hole position on the top side. This side (white board) were manufactured by hand for a third part person, so I must place the base support holes in the same place were the top side holes were placed. As you can observe, the holes are not well aligned.